requirement.txt
absl-py==0.6.1
astor==0.7.1
boto3==1.9.71
botocore==1.12.71
certifi==2018.11.29
Click==7.0
docutils==0.14
Flask==1.0.2
gast==0.2.0
gevent==1.3.7
greenlet==0.4.15
grpcio==1.16.1
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
mkl-fft==1.0.6
mkl-random==1.0.2
numpy==1.15.4
Pillow==5.3.0
protobuf==3.6.1
python-dateutil==2.7.5
s3transfer==0.1.13
scipy==1.1.0
six==1.12.0
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
